created kafka events for publishing user's rating

+++ rating-service/src/main/java/org/binitshrestha/ratingservice/dto/RatingCreateRequest.java
@@ -8,7 +8,7 @@ public record RatingCreateRequest(
     Long lawyerId,
     Long caseId,
     String ratingType,
-    int ratingValue,
+    double ratingValue,
     String comments
 ) {
 }
